requirements.txt
flask==2.0.1
anthropic==0.7.0
python-dotenv==0.19.0 